+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'sean@ertw.com'
 license          'MIT'
 description      'Resources for managing an F5 BigIP load balancer'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.4.4'
+version          '0.4.5'
 
 %w(ubuntu debian redhat centos suse opensuse opensuseleap scientific oracle amazon windows).each do |os|
   supports os
